+++ mlflow/tracking/context.py
@@ -1,0 +1,50 @@
+import os
+import sys
+import logging
+
+from mlflow.entities import SourceType
+
+
+_logger = logging.getLogger(__name__)
+
+
+def _get_main_file():
+    if len(sys.argv) > 0:
+        return sys.argv[0]
+    return None
+
+
+def _get_source_name():
+    main_file = _get_main_file()
+    if main_file is not None:
+        return main_file
+    return "<console>"
+
+
+def _get_git_commit(path):
+    try:
+        from git import Repo, InvalidGitRepositoryError, GitCommandNotFound, NoSuchPathError
+    except ImportError as e:
+        _logger.warning(
+            "Failed to import Git (the Git executable is probably not on your PATH),"
+            " so Git SHA is not available. Error: %s", e)
+        return None
+    try:
+        if os.path.isfile(path):
+            path = os.path.dirname(path)
+        repo = Repo(path, search_parent_directories=True)
+        commit = repo.head.commit.hexsha
+        return commit
+    except (InvalidGitRepositoryError, GitCommandNotFound, ValueError, NoSuchPathError):
+        return None
+
+
+def _get_source_version():
+    main_file = _get_main_file()
+    if main_file is not None:
+        return _get_git_commit(main_file)
+    return None
+
+
+def _get_source_type():
+    return SourceType.LOCAL
